% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietdataentry.R
\docType{package}
\name{dietdataentry}
\alias{dietdataentry}
\alias{dietdataentry-package}
\title{\pkg{dietdataentry}}
\description{
\pkg{dietdataentry}
}
\references{
\url{http://data.aad.gov.au/trophic}
}
