% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_dna_diet}
\alias{parse_dna_diet}
\title{Parse DNA diet data spreadsheet}
\usage{
parse_dna_diet(filename, dbh, worms_cache_directory = NULL,
  verbosity = 1, data_df, sources_df, refresh_worms_cache = FALSE)
}
\arguments{
\item{filename}{string: path to Excel file}

\item{dbh}{optional: database handle}

\item{worms_cache_directory}{string: path to cache directory for taxonomic data}

\item{verbosity}{numeric: 0 = silent, > 0 = give progress messages}

\item{data_df}{data.frame: pass the data data.frame directly rather than providing the filename}

\item{sources_df}{data.frame: pass the sources data.frame directly rather than providing the filename}

\item{refresh_worms_cache}{logical: if \code{TRUE}, refresh the taxonomic cache}
}
\value{
A list object with the parsed data
}
\description{
Parse DNA diet data spreadsheet
}
