% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_energetics}
\alias{parse_energetics}
\title{Parse energetics data spreadsheet}
\usage{
parse_energetics(
  filename,
  worms_cache_directory = NULL,
  verbosity = 1,
  refresh_worms_cache = FALSE
)
}
\arguments{
\item{filename}{string: path to Excel file}

\item{worms_cache_directory}{string: path to cache directory for taxonomic data}

\item{verbosity}{numeric: 0 = silent, > 0 = give progress messages}

\item{refresh_worms_cache}{logical: if \code{TRUE}, refresh the taxonomic cache}
}
\value{
A list object with the parsed data
}
\description{
Parse energetics data spreadsheet
}
