% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_sources}
\alias{parse_sources}
\title{Parse source details from spreadsheet}
\usage{
parse_sources(filename, verbosity = 1)
}
\arguments{
\item{filename}{string: path to Excel file}

\item{verbosity}{numeric: 0 = silent, > 0 = give progress messages}
}
\value{
A list object with the parsed data
}
\description{
Parse source details from spreadsheet
}
