% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{resolve_doi}
\alias{resolve_doi}
\title{Resolve a DOI}
\usage{
resolve_doi(doi)
}
\arguments{
\item{doi}{string: DOI to resolve}
}
\value{
The URL (as a string) that the DOI resolves
}
\description{
Resolve a DOI
}
\examples{
\dontrun{
  resolve_doi("10.26179/5d1aec22f41d5")
}
}
\references{
http://www.doi.org/
}
