% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_data.R
\name{so_append_data}
\alias{so_append_data}
\title{Append new SO diet data to existing tables}
\usage{
so_append_data(
  new_data,
  existing_data,
  existing_sources,
  records_type,
  action_if_existing = "stop",
  verbosity = 1
)
}
\arguments{
\item{new_data}{list: as returned by \code{\link[=parse_diet]{parse_diet()}}, \code{\link[=parse_isotopes]{parse_isotopes()}}, etc}

\item{existing_data}{data.frame: the existing diet (or isotopes, etc) data frame, as returned by \code{\link[sohungry:so_diet]{sohungry::so_diet()}}, \code{\link[sohungry:so_diet]{sohungry::so_isotopes()}}, etc}

\item{existing_sources}{data.frame: the existing sources data frame, as returned by \code{\link[sohungry:so_diet]{sohungry::so_sources()}}}

\item{records_type}{string: "diet", "dna_diet", "isotopes", "lipids", or "energetics"}

\item{action_if_existing}{string: what to do if we find one or more records in \code{new_data} that looks like an existing one in \code{existing_data}? Either "stop", "warn", or "ignore"}

\item{verbosity}{numeric: if greater than 0, show progress messages}
}
\value{
A list with updated records and sources data
}
\description{
Append new SO diet data to existing tables
}
