% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax-groups.R
\name{soki_ruleset}
\alias{soki_ruleset}
\title{Rule set for determining SOKI groups}
\usage{
soki_ruleset(worms_cache_directory = worms_cache_dir())
}
\arguments{
\item{worms_cache_directory}{string: path to a cache directory}
}
\value{
A function that takes the below parameters, and returns a string giving the SOKI group name
\itemize{
\item \code{worms_record} 1-row tibble: as returned by \code{\link[=search_worms]{search_worms()}}
\item \code{rule_type} string: "prey" or "predator"
\item \code{scientific} string: if \code{worms_record} not provided, the scientific name to search on
\item \code{aphia_id} integer: if \code{worms_record} not provided, the Aphia ID to search on
\item \code{worms_cache_directory} string: path to a cache directory
}
}
\description{
Rule set for determining SOKI groups
}
\examples{
\dontrun{
  rules <- soki_ruleset()
  rules(search_worms("Aptenodytes forsteri"), "predator")
}

}
