% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data}
\alias{validate_data}
\title{Validate data after parsing}
\usage{
validate_data(x, records_type = "guess", check_taxonomy = FALSE)
}
\arguments{
\item{x}{list: as returned by \code{\link[=parse_diet]{parse_diet()}}, \code{\link[=parse_isotopes]{parse_isotopes()}}, etc}

\item{records_type}{string: the data type "diet", "isotopes", "energetics", "dna_diet", "lipids", or "guess"}

\item{check_taxonomy}{logical: check the taxonomy against worrms?}
}
\value{
The input data.frame \code{x}, possibly with column types changed
}
\description{
Validate data after parsing
}
