% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax-worrms.R
\name{worms_cache_dir}
\alias{worms_cache_dir}
\title{Set the worms cache directory}
\usage{
worms_cache_dir(cache_directory)
}
\arguments{
\item{cache_directory}{string: path to a cache directory, or \code{NULL} to clear the current setting}
}
\description{
Set the worms cache directory
}
