% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax-worrms.R
\name{worms_common}
\alias{worms_common}
\title{Get common name for Aphia ID}
\usage{
worms_common(ids, cache_directory)
}
\arguments{
\item{ids}{numeric or string: AphiaID to match on}

\item{cache_directory}{string: path to a cache directory}
}
\value{
A tibble
}
\description{
Get common name for Aphia ID
}
