% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax-worrms.R
\name{worms_hierarchy}
\alias{worms_hierarchy}
\title{Get full hierarchy for Aphia ID}
\usage{
worms_hierarchy(id, cache_directory)
}
\arguments{
\item{id}{numeric or string: AphiaID to match on}

\item{cache_directory}{string: path to a cache directory}
}
\value{
A tibble
}
\description{
Get full hierarchy for Aphia ID
}
