% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietdataentry.R
\docType{package}
\name{dietdataentry}
\alias{dietdataentry}
\title{\pkg{dietdataentry}}
\description{
Functions for checking and entering data to the SCAR Southern Ocean Diet and Energetics Database
}
\references{
\url{https://scar.org/resources/southern-ocean-diet-energetics/}
}
